Partial fixes
