+++ src/app/core/lib/condition/condition.ts
@@ -7,9 +7,9 @@ import { fhirclient } from 'fhirclient/lib/types';
 import { MccCondition, MccConditionList, MccConditionSummary } from '../../types/mcc-types';
 import log from '../../utils/loglevel';
 import { getSupplementalDataClient } from '../goal/goal.util';
+import { fhirOptions, resourcesFrom } from '../../utils/fhir';
 import {
   notFoundResponse,
-  resourcesFrom,
   resourcesFromObject,
   transformToConditionSummary,
 } from './condition.util';
@@ -95,7 +95,7 @@ export const getSupplementalConditions = async (launchURL: string, sdsClient: Cl
             fhirHeaderRequestOption.url = 'Condition?subject=' + item2.resource.reference;
 
             // Fetch third-party goals
-            const response = await sdsClient.request(fhirHeaderRequestOption);
+            const response: fhirclient.JsonArray = await sdsClient.request(fhirHeaderRequestOption, fhirOptions);
 
             // Process third-party goals
             const thirdPartyGoals: Condition[] = resourcesFrom(response) as Condition[];
@@ -132,7 +132,7 @@ export function displayTransmitter(prov: Provenance | undefined): string | undef
 export const getSummaryConditions = async (sdsURL: string, authURL: string, sdsScope: string): Promise<MccConditionList> => {
 
   const client = await FHIR.oauth2.ready();
-  let sdsClient = await getSupplementalDataClient(client, sdsURL, authURL, sdsScope);
+  const sdsClient = await getSupplementalDataClient(client, sdsURL, authURL, sdsScope);
 
   const activeConcerns: MccConditionSummary[] = [];
   const activeConditions: MccConditionSummary[] = [];
@@ -142,12 +142,12 @@ export const getSummaryConditions = async (sdsURL: string, authURL: string, sdsS
   const queryPath1 = `Condition?category=http%3A%2F%2Fterminology.hl7.org%2FCodeSystem%2Fcondition-category%7Cproblem-list-item&_revinclude=Provenance:target`;
   const queryPath2 = `Condition?category=http%3A%2F%2Fhl7.org%2Ffhir%2Fus%2Fcore%2FCodeSystem%2Fcondition-category%7Chealth-concern&_revinclude=Provenance:target`;
 
-  const conditionRequest1: fhirclient.JsonObject = await client.patient.request(queryPath1);
-  const conditionRequest2: fhirclient.JsonObject = await client.patient.request(queryPath2);
-
+  const conditionRequest1: fhirclient.JsonArray = await client.patient.request(queryPath1, fhirOptions);
+  const conditionRequest2: fhirclient.JsonArray = await client.patient.request(queryPath2, fhirOptions);
+  
   let sdsfilteredConditions2: MccCondition[] = [];
   if (sdsClient) {
-    const sdsconditionRequest2: fhirclient.JsonObject = await sdsClient.patient.request(queryPath2);
+    const sdsconditionRequest2: fhirclient.JsonArray = await sdsClient.patient.request(queryPath2, fhirOptions);
     sdsfilteredConditions2 = resourcesFrom(sdsconditionRequest2) as MccCondition[];
   }
 
@@ -255,8 +255,8 @@ export const getConditions = async (): Promise<MccCondition[]> => {
   const client = await FHIR.oauth2.ready();
 
   const queryPath = `Condition`;
-  const conditionRequest: fhirclient.JsonObject = await client.patient.request(
-    queryPath
+  const conditionRequest: fhirclient.JsonArray = await client.patient.request(
+    queryPath, fhirOptions
   );
 
   log.debug({ condRequest: conditionRequest })
@@ -266,7 +266,7 @@ export const getConditions = async (): Promise<MccCondition[]> => {
   ) as MccCondition[];
 
   log.info(
-    `getConditions - successful`
+    `getConditions - successful with length ${filteredConditions.length}`
   );
   log.debug({ serviceName: 'getConditions', result: filteredConditions });
   return filteredConditions;
