+++ src/app/core/lib/contact/contact.util.ts
@@ -4,55 +4,6 @@ import { fhirclient } from 'fhirclient/lib/types';
 import { MccPatientContact } from '../../types/mcc-types';
 import { getConceptDisplayString } from '../goal/goal.util';
 
-// import log from '../../utils/loglevel';
-
-export const fhirOptions: fhirclient.FhirOptions = {
-  pageLimit: 0,
-};
-
-export const notFoundResponse = (code?: string) => ({
-  code,
-  status: 'notfound',
-  value: {
-    stringValue: 'No Data Available',
-    valueType: 'string',
-  },
-});
-
-export const resourcesFromObject = (
-  response: fhirclient.JsonObject
-): Resource => {
-  const entry: fhirclient.JsonObject = response?.entry[0];
-
-  const resource: any = entry?.resource;
-
-  if (resource.resourceType === 'OperationOutcome') {
-    return {} as any;
-  }
-
-  return resource;
-};
-
-export const resourcesFrom = (response: fhirclient.JsonObject): Resource[] => {
-  const entries = (response[0] as fhirclient.JsonObject)?.entry as [fhirclient.JsonObject];
-  return entries?.map((entry: fhirclient.JsonObject) => entry.resource as any)
-    .filter((resource: Resource) => resource.resourceType !== 'OperationOutcome');
-};
-
-export const resourcesFrom2 = (response: fhirclient.JsonArray): Resource[] => {
-  const firstEntries = response[0] as fhirclient.JsonObject;
-  const entries: fhirclient.JsonObject[] = firstEntries?.entry
-    ? (firstEntries.entry as [fhirclient.JsonObject])
-    : [];
-
-  return entries
-    .map((entry: fhirclient.JsonObject) => entry?.resource as any)
-    .filter(
-      (resource: Resource) => resource.resourceType !== 'OperationOutcome'
-    );
-};
-
-
 function resolve(ref?: Reference, members?: Map<string, Practitioner>) {
   let resourceID: string | undefined = ref?.reference?.split('/').reverse()?.[0]
   return members?.get(resourceID ?? 'missing id')
