+++ src/app/core/lib/observation/observation.ts
@@ -7,12 +7,12 @@ import { getAllCodes, getCodesFromValueSetCode } from '../../query/json';
 import type { ObservationMode } from '../../types';
 import { legacy_MccObservationCollection, MccObservation } from '../../types/mcc-types';
 import log from '../../utils/loglevel';
+import { resourcesFrom } from '../../utils/fhir';
 
 import {
   convertToObservationCollection,
   fhirOptions,
   notFoundResponse,
-  resourcesFrom,
 } from './observation.util';
 
 export const getObservation = async (code: string): Promise<MccObservation> => {
