+++ src/app/core/resources/SocialConcerns.json
