+++ src/app/health-and-social-concerns/health-and-social-concerns.component.html
@@ -7,7 +7,7 @@
   </section>
   <section>
     <div class="table-container">
-      <app-social-concern-panel  *ngIf="featureToggling.healthAndSocialConcerns.activeDiagnoses && getActiveIsReady()"></app-social-concern-panel>
+      <app-social-concern-panel  *ngIf="featureToggling.healthAndSocialConcerns.socialConcerns && getActiveIsReady()"></app-social-concern-panel>
     </div>
   </section>
   <section>
